refactor(platform): fix test file locations and testing convention violations
- move zero-about-page tests to views/zero-page/__tests__/ (co-located with component)
- move zero-team-detail-page tests to views/team-page/__tests__/ (co-located with component)
- split mixed-component test file into two focused files
- replace UI text content assertions (AP-8) with role-based structural assertions
- replace signal state assertion (AP-7) in back-button test with DOM visibility assertion
